Extract SDR output writers to output/sdr/ subpackage
Move apply_excel_formatting, write_csv/json/html/markdown_output,
ExcelFormatCache, and write_excel_output (~930 LOC) from generator.py
to output/sdr/__init__.py. Update lazy forwarders in output/writers/
and output/excel.py to target the new module directly.
