+++ src/cja_auto_sdr/output/writers/html.py
@@ -2,4 +2,4 @@ __all__ = ["write_html_output"]
 
 from cja_auto_sdr.core.lazy import make_getattr
 
-__getattr__ = make_getattr(__name__, __all__, target_module="cja_auto_sdr.generator")
+__getattr__ = make_getattr(__name__, __all__, target_module="cja_auto_sdr.output.sdr")
